+++ python/ray/llm/_internal/serve/deployments/llm/multiplex/utils.py
@@ -8,14 +8,12 @@ from fastapi import HTTPException
 from filelock import FileLock
 
 from ray.llm._internal.serve.observability.logging import get_logger
-from ray.llm._internal.serve.deployments.utils.cloud_utils import (
+from ray.llm._internal.common.utils.cloud_utils import (
     CloudFileSystem,
+    LoraMirrorConfig,
     remote_object_cache,
 )
-from ray.llm._internal.serve.configs.server_models import (
-    LLMConfig,
-    LoraMirrorConfig,
-)
+from ray.llm._internal.serve.configs.server_models import LLMConfig
 from ray.llm._internal.serve.configs.constants import (
     CLOUD_OBJECT_MISSING_EXPIRE_S,
     CLOUD_OBJECT_EXISTS_EXPIRE_S,
