+++ python/ray/llm/_internal/serve/deployments/llm/vllm/vllm_models.py
@@ -13,9 +13,9 @@ from ray.llm._internal.utils import try_import
 
 from ray.llm._internal.serve.observability.logging import get_logger
 from ray.llm._internal.common.base_pydantic import BaseModelExtended
+from ray.llm._internal.common.utils.cloud_utils import CloudMirrorConfig
 from ray.llm._internal.serve.configs.server_models import (
     DiskMultiplexConfig,
-    CloudMirrorConfig,
     GenerationRequest,
     GPUType,
     LLMConfig,
