+++ python/ray/llm/_internal/serve/deployments/utils/node_initialization_utils.py
@@ -11,14 +11,10 @@ from ray.llm._internal.utils import try_import
 
 from ray.llm._internal.serve.observability.logging import get_logger
 
-from ray.llm._internal.serve.deployments.utils.downloader_utils import (
-    CloudModelDownloader,
-)
+from ray.llm._internal.common.utils.cloud_utils import CloudMirrorConfig
+from ray.llm._internal.common.utils.download_utils import CloudModelDownloader
 from ray.llm._internal.serve.deployments.llm.vllm.vllm_models import VLLMEngineConfig
-from ray.llm._internal.serve.configs.server_models import (
-    CloudMirrorConfig,
-    LLMConfig,
-)
+from ray.llm._internal.serve.configs.server_models import LLMConfig
 from ray.llm._internal.serve.deployments.utils.server_utils import make_async
 
 torch = try_import("torch")
